style reorg, json-only samples with client side template implementations
