+++ src/store.ts
@@ -4,7 +4,64 @@ import { JSONSchema7 } from "json-schema";
 import { create } from "zustand";
 import { AppState } from "./types/store";
 import { getSamplesList, getSampleByName } from "./api";
+import { templateComponents } from "@/templates";
 
+// List of UI keys that should be mapped to template components
+const templateKeys = ["ui:ObjectFieldTemplate", "ui:field", "ui:ArrayFieldTemplate", "ui:widget"];
+
+// Process schema to handle large enums
+function processSchema(schema: any): any {
+  if (!schema || typeof schema !== 'object') return schema;
+
+  // Create a deep copy to avoid modifying the original
+  const result = JSON.parse(JSON.stringify(schema));
+
+  // Process definitions for large enums
+  if (result.definitions) {
+    Object.keys(result.definitions).forEach(key => {
+      const def = result.definitions[key];
+      if (def.type === 'string' && def.enumSize && typeof def.enumSize === 'number') {
+        // Generate the enum values
+        def.enum = Array.from({ length: def.enumSize }, (_, i) => `option #${i}`);
+        // Remove the marker
+        delete def.enumSize;
+      }
+    });
+  }
+
+  // Process other properties recursively
+  Object.keys(result).forEach(key => {
+    if (result[key] && typeof result[key] === 'object') {
+      result[key] = processSchema(result[key]);
+    }
+  });
+
+  return result;
+}
+
+// Helper function to recursively process uiSchema and replace string references with components
+function processUiSchema(obj: any): any {
+  if (!obj || typeof obj !== 'object') return obj;
+
+  // Create a copy to modify
+  const result = Array.isArray(obj) ? [...obj] : {...obj};
+
+  // Process all properties
+  for (const key of Object.keys(result)) {
+    if (templateKeys.includes(key) && typeof result[key] === 'string') {
+      const componentName = result[key];
+      if (componentName in templateComponents) {
+        // Replace string reference with actual component
+        result[key] = templateComponents[componentName];
+      }
+    } else if (result[key] && typeof result[key] === 'object') {
+      // Recursively process nested objects
+      result[key] = processUiSchema(result[key]);
+    }
+  }
+
+  return result;
+}
 
 export const useStore = create<AppState>((set, get) => ({
   schema: {} as JSONSchema7 | RJSFSchema,
@@ -15,9 +72,7 @@ export const useStore = create<AppState>((set, get) => ({
   error: null,
   availableSamples: [],
 
-  // Fetch list of available samples
   fetchSamples: async () => {
-    // If we already have samples, don't fetch again
     if (get().availableSamples.length > 0 || get().loading) {
       return;
     }
@@ -31,7 +86,6 @@ export const useStore = create<AppState>((set, get) => ({
         loading: false
       });
 
-      // If we have samples but no selected sample, select the second one
       if (samplesList.length > 0 && !get().label) {
         const firstSample = samplesList[1];
         set({ label: firstSample });
@@ -45,15 +99,23 @@ export const useStore = create<AppState>((set, get) => ({
     }
   },
 
-  // Fetch a specific sample by name
   fetchSample: async (sampleName: string) => {
     try {
       set({ loading: true, error: null });
       const sample = await getSampleByName(sampleName);
 
+      // Process schema to handle large enums
+      const processedSchema = processSchema(sample.schema);
+
+      // Process uiSchema to replace string references with actual components
+      let processedUiSchema = {};
+      if (sample.uiSchema && typeof sample.uiSchema === 'object') {
+        processedUiSchema = processUiSchema(sample.uiSchema);
+      }
+
       set({
-        schema: sample.schema as JSONSchema7 | RJSFSchema,
-        uiSchema: sample.uiSchema as object,
+        schema: processedSchema as JSONSchema7 | RJSFSchema,
+        uiSchema: processedUiSchema,
         formData: sample.formData as object,
         loading: false
       });
@@ -65,13 +127,11 @@ export const useStore = create<AppState>((set, get) => ({
     }
   },
 
-  // Set the current sample label and fetch the sample data
   setLabel: (label: string) => {
     set({ label });
     get().fetchSample(label);
   },
 
-  // Update the schema, uiSchema, and formData
   updateSchema: (schema: JSONSchema7 | RJSFSchema) => {
     set({ schema });
   },
