+++ src/types/store.ts
@@ -1,6 +1,6 @@
 // src/types/store.ts
-import { RJSFSchema } from "@rjsf/utils"
-import { JSONSchema7 } from "json-schema"
+import { RJSFSchema } from "@rjsf/utils";
+import { JSONSchema7 } from "json-schema";
 
 export interface AppState {
   schema: JSONSchema7 | RJSFSchema;
@@ -8,7 +8,7 @@ export interface AppState {
   formData: object;
   label: string;
   loading: boolean;
-  error: null | string;
+  error: string | null;
   availableSamples: string[];
   fetchSamples: () => Promise<void>;
   fetchSample: (sampleName: string) => Promise<void>;
